adding build changes

+++ hadoop-1.2.1/src/hdfs/org/apache/hadoop/mdfs/comm/network/edu/tamu/lenss/mdfs/comm/DeleteFileHandler.java
@@ -7,17 +7,19 @@ import java.util.Set;
 
 import adhoc.etc.IOUtilities;
 import edu.tamu.lenss.mdfs.Constants;
-import edu.tamu.lenss.mdfs.MDFSDirectory;
+//import edu.tamu.lenss.mdfs.MDFSDirectory;
 import edu.tamu.lenss.mdfs.models.DeleteFile;
 import edu.tamu.lenss.mdfs.models.MDFSFileInfo;
 import edu.tamu.lenss.mdfs.utils.AndroidIOUtils;
+
+import org.apache.hadoop.mdfs.protocol.MDFSDirectoryProtocol;
 
 public class DeleteFileHandler {
 	
 	public DeleteFileHandler(){
 	}
 	
-	public void processPacket(DeleteFile delete){
+	public void processPacket(DeleteFile delete,boolean local){
 		if(delete.getFileIds() == null || delete.getFileNames() == null){
 			return;
 		}
@@ -25,7 +27,7 @@ public class DeleteFileHandler {
 		if(delete.getFileIds().size() != delete.getFileNames().size()){
 			return;
 		}
-		new DeleteFiileThread(delete.getFileNames(), delete.getFileIds()).start();
+		new DeleteFiileThread(delete.getFileNames(), delete.getFileIds(),local).start();
 	}
 	
 	protected void deleteFiles(Set<Long> mergeFileIds){
@@ -42,15 +44,17 @@ public class DeleteFileHandler {
 		private Set<Long> fileIds;
 		private List<String> fNames; 
 		private List<Long> fIds;
+		private boolean local;
 		
 		private DeleteFiileThread(Set<Long> mergeFileIds){
 			fileIds = mergeFileIds;
 		}
 		
 
-		private DeleteFiileThread(List<String> fNames, List<Long> fIds){
+		private DeleteFiileThread(List<String> fNames, List<Long> fIds,boolean local){
 			this.fIds = fIds;
 			this.fNames = fNames;
+			this.local=local;
 		}
 		
 		@Override
@@ -66,12 +70,14 @@ public class DeleteFileHandler {
 				fName = fNames.get(i);
 				
 				// Clean up the MDFSDirectory
-				MDFSDirectory directory = ServiceHelper.getInstance().getDirectory();
-				directory.removeDecryptedFile(fileId);
-				directory.removeEncryptedFile(fileId);
-				directory.removeFile(fileId);
-				directory.removeFileFragment(fileId);
-				directory.removeKeyFragment(fileId);
+				if(local){
+					MDFSDirectoryProtocol directory = ServiceHelper.getInstance().getDirectory();
+					directory.removeDecryptedFile(fileId);
+					directory.removeEncryptedFile(fileId);
+					directory.removeFile(fileId);
+					directory.removeFileFragment(fileId,ServiceHelper.getInstance().getMyNode().getNodeId());
+					directory.removeKeyFragment(fileId,ServiceHelper.getInstance().getMyNode().getNodeId());
+				}
 				
 				// Remove Encrypted File
 				File file = new File(rootDir, "encrypted/" + MDFSFileInfo.getDirName(fName, fileId));
@@ -87,11 +93,13 @@ public class DeleteFileHandler {
 				File fileDir = new File(rootDir, MDFSFileInfo.getDirName(fName, fileId));
 				System.out.println(" File  to be deleted "+fileDir.getName());
 					
-				try {
-					IOUtilities.cleanCache(fileDir, 0);
-					fileDir.delete();
-				} catch (FileNotFoundException e) {
-					e.printStackTrace();
+				if(local){
+					try {
+						IOUtilities.cleanCache(fileDir, 0);
+						fileDir.delete();
+					} catch (FileNotFoundException e) {
+						e.printStackTrace();
+					}
 				}				
 				
 			}
